% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bowl18.R
\name{bowl18_min_freq_dist}
\alias{bowl18_min_freq_dist}
\alias{bowl18_min_freq_dist.default}
\alias{bowl18_min_freq_dist.fr_chord}
\title{Minimum frequency distance}
\usage{
bowl18_min_freq_dist(x)

\method{bowl18_min_freq_dist}{default}(x)

\method{bowl18_min_freq_dist}{fr_chord}(x)
}
\arguments{
\item{x}{Chord to analyse.
The default method assumes that the chord is expressed
as a numeric vector of frequencies.
Representations from the hrep package
(e.g. \code{\link[hrep]{pi_chord}()})
can be used to analyse chords expressed as MIDI note numbers.}
}
\value{
(Numeric scalar)
The minimum distance between the fundamental frequencies of the chord,
in Hz.
}
\description{
This function returns the minimum distance between
the fundamental frequencies of a chord,
after Bowling et al. (2018).
It makes no assumptions about the chord's tuning.
}
\examples{
bowl18_min_freq_dist(c(220, 440, 560)) # processed as frequencies
bowl18_min_freq_dist(hrep::fr_chord(c(220, 440, 560))) # same as above
bowl18_min_freq_dist(hrep::pi_chord(c(60, 64, 67))) # MIDI note numbers
}
\references{
\insertRef{Bowling2018}{incon}
}
