% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{complex_sonor}
\alias{complex_sonor}
\alias{complex_sonor.parn94}
\alias{complex_sonor.default}
\title{Get complex sonorousness}
\usage{
complex_sonor(x, k_c = parn94_params()$k_c, ...)

\method{complex_sonor}{parn94}(x, k_c = parn94_params()$k_c, ...)

\method{complex_sonor}{default}(x, k_c = parn94_params()$k_c, ...)
}
\arguments{
\item{x}{Object to analyse.}

\item{k_c}{Parncutt & Strasburger (1994) set this to 0.2 (p. 105)}

\item{...}{Further parameters to pass to \code{\link{parn94}()}.}
}
\value{
Complex sonorousness, a numeric scalar.
}
\description{
Computes the complex sonorousness of a sound, after
\insertCite{Parncutt1994;textual}{incon}.
}
\references{
\insertAllCited{}
}
