% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-corpdiss.R
\name{corpus_dissonance_table}
\alias{corpus_dissonance_table}
\alias{corpus_dissonance_table.corpus}
\title{Corpus dissonance table}
\usage{
corpus_dissonance_table(x, type = "pc_chord_type", add = 1L)

\method{corpus_dissonance_table}{corpus}(x, type = "pc_chord_type", add = 1L)
}
\arguments{
\item{x}{Corpus to analyse, as created by \code{hrep::\link[hrep]{corpus}}.}

\item{type}{(Scalar character) Representation to which chords should be coerced
before counting.}

\item{add}{(Scalar numeric) Number to add to each count before computing probabilities.
This is useful for ensuring that chord probabilities exceed zero.}
}
\value{
Returns an object of class \code{corpus_dissonance_table},
a \link[tibble]{tibble} where each row corresponds to a
different pitch-class chord type (i.e. an object of class \code{pc_chord_type}),
with the mapping between integer indices and chord types defined by
the \code{hrep} package.
This tibble contains the following columns:
\item{count}{The number of times the chord type was observed in the corpus.}
\item{prob}{The inferred probability of that chord type.}
\item{log_prob}{The inferred log probability of that chord type.}
}
\description{
Derives sufficient information from a corpus to formulate
a corpus-based dissonance model.
}
