% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-har18.R
\name{cosine_similarity}
\alias{cosine_similarity}
\title{Cosine similarity}
\usage{
cosine_similarity(x, y)
}
\arguments{
\item{x}{Numeric vector 1.}

\item{y}{Numeric vector 2.}
}
\value{
Cosine similarity, as a numeric scalar.
}
\description{
Computes the cosine similarity between two numeric vectors.
}
