% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-wang13.R
\name{demo_wang}
\alias{demo_wang}
\title{Wang et al. (2013) interactive demo}
\usage{
demo_wang(audio = TRUE)
}
\arguments{
\item{audio}{(Scalar logical) Whether to enable playback controls
(currently the playback controls don't work when the app is hosted
on a remote server).}
}
\description{
Launches an interactive demo of Wang et al.'s (2013) roughness model.
This function requires a few additional packages to be installed;
you will be notified if any of these packages are missing
once you run \code{demo_wang()}.
}
\note{
The demo takes the form of an Shiny app
(\url{https://shiny.rstudio.com/}).
}
\references{
\insertRef{Wang2013}{incon}
}
