% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{dyad_roughness_seth}
\alias{dyad_roughness_seth}
\title{Dyad roughness (Sethares)}
\usage{
dyad_roughness_seth(
  f1,
  f2,
  a1,
  a2,
  ensure_f1_is_less_than_f2 = TRUE,
  min_amplitude = TRUE,
  a = 3.5,
  b = 5.75,
  s1 = 0.021,
  s2 = 19,
  d_star = 0.24
)
}
\arguments{
\item{f1}{Frequency of tone 1 (Hz) (numeric vector).}

\item{f2}{Frequency of tone 2 (Hz) (numeric vector). Must be greater than \code{f1}.}

\item{a1}{amplitude of tone 1 (numeric vector).}

\item{a2}{amplitude of tone 2 (numeric vector).}

\item{ensure_f1_is_less_than_f2}{If \code{FALSE}, assumes that \code{f1 < f2}.}

\item{min_amplitude}{If \code{TRUE},
roughness is considered to be proportional to
the minimum amplitude of each pair of partials,
rather than the product of their amplitudes.
The default (\code{TRUE}) corresponds to the algorithm as updated by
\insertCite{Sethares2005;textual}{incon} and
\insertCite{Weisser2013;textual}{incon}.
Set to \code{FALSE} to recover the original algorithm from
\insertCite{Sethares1993;textual}{incon}.}

\item{a}{Numeric scalar parameter, optimised to 3.5 (default) in Sethares (1993).}

\item{b}{Numeric scalar parameter, optimised to 5.75 (default) in Sethares (1993).}

\item{s1}{Numeric scalar parameter from Sethares (1993).}

\item{s2}{Numeric scalar parameter from Sethares (1993).}

\item{d_star}{Numeric scalar parameter from Sethares (1993).}
}
\value{
Numeric vector of roughnesses.
}
\description{
Gets the roughness of a dyad according to the model of Sethares (1993).
}
\references{
\insertAllCited{}
}
