% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{dyad_roughness_vass}
\alias{dyad_roughness_vass}
\title{Dyad roughness (Vassilakis)}
\usage{
dyad_roughness_vass(f1, f2, a1, a2)
}
\arguments{
\item{f1}{Frequency of tone 1 (Hz) (numeric vector).}

\item{f2}{Frequency of tone 2 (Hz) (numeric vector).}

\item{a1}{amplitude of tone 1 (numeric vector).}

\item{a2}{amplitude of tone 2 (numeric vector).}
}
\value{
Numeric vector of roughnesses.
}
\description{
Gets the roughness of a dyad according to the model of
\insertCite{Vassilakis2001;textual}{incon}
\insertCite{Villegas2010;textual}{incon}
}
\note{
The function is vectorised over all inputs.
}
\references{
\insertAllCited{}
}
