% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{get_free_field_threshold}
\alias{get_free_field_threshold}
\title{Get free field threshold}
\usage{
get_free_field_threshold(kHz)
}
\arguments{
\item{kHz}{Numeric vector of frequencies in kHz.}
}
\value{
Numeric vector of corresponding free-field thresholds in dB SPL.
}
\description{
Returns the free-field threshold (dB SPL) of hearing in quiet for pure tones
of given frequencies.
This is the minimum sound level at which a pure tone at that frequency
will be heard.
Corresponds to Equation 2 in \insertCite{Parncutt1994;textual}{incon}.
}
\references{
\insertAllCited{}
}
