% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{get_overall_masking_level}
\alias{get_overall_masking_level}
\title{Get overall masking level}
\usage{
get_overall_masking_level(auditory_level, pure_tone_height, k_m)
}
\arguments{
\item{auditory_level}{Numeric vector of auditory levels}

\item{pure_tone_height}{Numeric vector of pure tone heights}

\item{k_m}{See \code{\link{parn94_params}()}.}
}
\value{
Numeric vector of overall masking levels (dB).
}
\description{
Returns overall masking levels for a set of pure tones
that are assumed to be playing simultaneously.
Corresponds to Parncutt & Strasburger (1994) Equation 5.
}
\references{
\insertAllCited{}
}
