% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{get_partial_masking_level}
\alias{get_partial_masking_level}
\title{Get partial masking level}
\usage{
get_partial_masking_level(
  masker_auditory_level,
  masker_pure_tone_height,
  maskee_auditory_level,
  maskee_pure_tone_height,
  k_m
)
}
\arguments{
\item{masker_auditory_level}{Numeric vector of masker auditory levels.}

\item{masker_pure_tone_height}{Numeric vector of masker pure tone heights.}

\item{maskee_auditory_level}{Numeric vector of maskee auditory levels.}

\item{maskee_pure_tone_height}{Numeric vector of maskee pure tone heights.}

\item{k_m}{Parameter \code{k_m} in \insertCite{Parncutt1994;textual}{parn94}.
represents the masking pattern gradient for a pure tone,
with units of dB per critical band.
Parncutt & Strasburger use a value of 12 in their examples,
but imply that 12-18 is a typical range of values for this parameter.}
}
\value{
Matrix where element [i, j] gives the level of masking
for masker j on maskee i.
}
\description{
Returns the effective reduction in dB of the audible level
of a masked pure tone (maskee) on account of a masking pure tone (masker).
Equation 4 in \insertCite{Parncutt1994;textual}{parn94}.
}
\references{
\insertAllCited{}
}
