% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{get_pure_tone_audibility}
\alias{get_pure_tone_audibility}
\title{Get audibility}
\usage{
get_pure_tone_audibility(pure_tone_audible_level, al_0)
}
\arguments{
\item{pure_tone_audible_level}{Numeric vector of audible levels (dB).}

\item{al_0}{constant (see Equation 7 of Parncutt & Strasburger (1994)).}
}
\value{
Numeric vector of pure tone audibilities.
}
\description{
Returns the audibility of a set of pure tone components as a
function of their audible levels.
Corresponds to Equation 7 of \insertCite{Parncutt1994;textual}{incon}.
}
\references{
\insertAllCited{}
}
