% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{get_pure_tone_height}
\alias{get_pure_tone_height}
\title{Get pure-tone height}
\usage{
get_pure_tone_height(kHz)
}
\arguments{
\item{kHz}{Numeric vector of frequencies in kHz.}
}
\value{
Numeric vector of corresponding pure-tone heights,
with units of equivalent rectangular bandwidths (ERBs).
}
\description{
Returns the pure-tone heights (a.k.a. critical-band rates)
of pure tones of given frequencies.
Equation 3 in \insertCite{Parncutt1994;textual}{parn94}.
}
\references{
\insertAllCited{}
}
