% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bowl18.R
\name{gill09_harmonicity}
\alias{gill09_harmonicity}
\title{Harmonicity}
\usage{
gill09_harmonicity(x, tonic = 0L)
}
\arguments{
\item{x}{Chord to analyse,
expressed as an integerish vector of MIDI note numbers,
or more generally as any valid input to \code{\link[hrep:pi_chord]{pi_chord}}.}

\item{tonic}{(Integerish scalar, default = 0L)
Tonic to use when defining the just-tuned scale.}
}
\value{
(Numeric scalar)
The chord's harmonicity, defined as the proportion of harmonics
of the chord's fundamental frequency that coincide with the
harmonics of the chord tones.
}
\description{
Computes the harmonicity of a chord using the
percentage similarity measure of Gill & Purves (2009).
}
\details{
This percentage similarity measure corresponds to the percentage of the harmonics
that the chord holds in common with a harmonic series rooted on
the chord's fundamental frequency.

While \insertCite{Gill2009;textual}{incon} originally presented this measure
to quantify the harmonicity of two-note chords (dyads)
\insertCite{Bowling2018;textual}{incon} subsequently demonstrated the application
of this measure to chords of arbitrary sizes.
}
\note{
The algorithm assumes that chord pitches are precisely
tuned to the just-tuned scale provided by Bowling et al. (2018).
This scale can be found at \code{\link{rational_scale}}.
}
\examples{
gill09_harmonicity(c(60, 64, 67))
gill09_harmonicity("60 64 67")
gill09_harmonicity(hrep::pi_chord(c(60, 64, 67)))
}
\references{
\insertRef{Bowling2018}{incon}
}
