% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-har-2019.R
\name{har_19_composite}
\alias{har_19_composite}
\title{Composite consonance model}
\usage{
har_19_composite(x, chord_size = FALSE, num_harmonics = 11L, roll_off = 1)
}
\arguments{
\item{x}{Chord to analyse; passed to \code{\link{incon}}.
All chord pitches must be integer-valued.}

\item{chord_size}{(Boolean) Whether to include the chord size effect
(see \strong{Details}).}

\item{num_harmonics}{(Integerish scalar)
Number of harmonics in each tone; passed to \code{\link{incon}}.}

\item{roll_off}{(Numeric scalar)
Harmonic roll-off rate; passed to \code{\link{incon}}.}
}
\value{
Estimated consonance, as a numeric scalar.
}
\description{
This function computes the consonance of a chord
according to Harrison & Pearce's (in preparation)
composite consonance model.
}
\details{
The model combines several sub-models:
\itemize{
\item \code{hutch_78_roughness},
the roughness model of \insertCite{Hutchinson1978;textual}{dycon}
(see \code{dycon::\link[dycon]{roughness_hutch}});
\item \code{har_18_harmonicity},
the harmonicity model of \insertCite{Harrison2018;textual}{har18}
(see \code{har18::\link[har18]{pc_harmonicity}});
\item \code{har_19_corpus}:
a corpus-based model of cultural familiarity
(Harrison & Pearce, in preparation)
(see \code{corpdiss::\link[corpdiss]{corpus_dissonance}}).
}

This model uses the regression coefficients
provided in \code{\link{har_19_composite_coef}}, with one caveat:
by default, the chord size effect is disabled,
because it's thought that this effect came from a confound
in the perceptual data of \insertCite{Bowling2018;textual}{incon}.
}
\references{
\insertAllCited{}
}
