% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/har-2019.R
\docType{data}
\name{har_19_composite_coef}
\alias{har_19_composite_coef}
\title{Regression coefficients}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 5 rows and 2 columns.}
\usage{
har_19_composite_coef
}
\description{
Tabulates regression coefficients for Harrison & Pearce's (in preparation)
composite consonance model.
}
\keyword{data}
