% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-huron-1994.R
\name{huron_1994}
\alias{huron_1994}
\title{Dyadic consonance (Huron 1994)}
\usage{
huron_1994(x)
}
\arguments{
\item{x}{Chord to evaluate, will be coerced to a
pitch-class set (\code{\link[hrep]{pc_set}}).}
}
\description{
Computes the aggregate dyadic consonance of a chord following
Huron (1994).
}
