% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-huron-1994.R
\docType{data}
\name{huron_1994_weights}
\alias{huron_1994_weights}
\title{Dyadic consonance weights (Huron 1994)}
\format{
An object of class \code{numeric} of length 6.
}
\usage{
huron_1994_weights
}
\description{
A vector of the weights used in Huron's (1994) dyadic consonance model.
}
\keyword{datasets}
