% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{hutch_cbw}
\alias{hutch_cbw}
\title{Critical bandwidth (Hutchison & Knopoff)}
\usage{
hutch_cbw(f1, f2)
}
\arguments{
\item{f1}{(Numeric vector) Frequency 1, Hz}

\item{f2}{(Numeric vector) Frequency 2, Hz}
}
\value{
(Numeric vector) Critical bandwidths.
}
\description{
Calculates the critical bandwidth given pairs of frequencies
\code{f1} and \code{f2},
according to the model of \insertCite{Hutchinson1978;textual}{incon}.
}
\references{
\insertAllCited{}
}
