% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{hutch_dissonance_function}
\alias{hutch_dissonance_function}
\title{Get dissonance contribution}
\usage{
hutch_dissonance_function(f1, f2, cbw_cut_off = 1.2, a = 0.25, b = 2)
}
\arguments{
\item{f1}{(Numeric vector) Frequency 1, Hz}

\item{f2}{(Numeric vector) Frequency 2, Hz}

\item{cbw_cut_off}{Parameter passed to \code{\link{hutch_g}()}.}

\item{a}{Parameter passed to \code{\link{hutch_g}()}.}

\item{b}{Parameter passed to \code{\link{hutch_g}()}.}
}
\value{
Numeric vector of dissonance contributions.
}
\description{
Computes the dissonance contribution of a pair of pure tones.
}
