% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{hutch_g}
\alias{hutch_g}
\title{Dissonance factor (Hutchinson & Knopoff)}
\usage{
hutch_g(y, cbw_cut_off = 1.2, a = 0.25, b = 2)
}
\arguments{
\item{y}{(Numeric vector) Frequency distance in units of critical bandwidths.}

\item{cbw_cut_off}{(Numeric scalar)
If not \code{NULL}, then should be a number
corresponding to the variable CBWcutoff in Mashinter's own implementation.
If \code{y >= cbw_cut_off}, then the dissonance factor will be approximated as 0.
Setting \code{cbw_cut_off} to 1.2 is necessary for replicating Mashinter's results.
A cut-off of 1.2 was also used by \insertCite{Bigand1996;textual}{incon}.}

\item{a}{(Numeric scalar, default = 0.25)
Parameter from \insertCite{Mashinter2006;textual}{incon}.}

\item{b}{(Numeric scalar, default = 2)
Parameter from \insertCite{Mashinter2006;textual}{incon}.}
}
\value{
(Numeric vector) Dissonance factors.
}
\description{
Computes dissonance factors given frequency distances in units of
critical bandwidths, after Mashinter's implementation of
Hutchinson & Knopoff's model
\insertCite{Mashinter2006,Hutchinson1978}{incon}.
This function corresponds to an approximation of the
look-up table in \insertCite{Plomp1965;textual}{incon}.
}
\references{
\insertAllCited{}
}
