% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{hutch_visualise}
\alias{hutch_visualise}
\title{Visualise}
\usage{
hutch_visualise(
  x,
  cbw_cut_off = 1.2,
  a = 0.25,
  b = 2,
  label = "Roughness",
  amplitude_breaks = c(0, 1),
  colour_limits = c(0, 3),
  colour_low = "darkblue",
  colour_high = "red",
  theme = hutch_visualise_theme,
  ...
)
}
\arguments{
\item{x}{Passed to \code{\link{roughness_hutch}}.}

\item{cbw_cut_off}{Passed to \code{\link{roughness_hutch}}.}

\item{a}{Passed to \code{\link{roughness_hutch}}.}

\item{b}{Passed to \code{\link{roughness_hutch}}.}

\item{label}{(Character scalar) x-axis label.}

\item{amplitude_breaks}{Numeric vector of tick locations for the y-axis.}

\item{colour_limits}{Defines the limits of the roughness scale.}

\item{colour_low}{Colour to use for the lowest roughness.}

\item{colour_high}{Colour to use for the highest roughness.}

\item{theme}{\code{\link[ggplot2]{ggplot}} theme to use.}

\item{...}{Passed to \code{\link[hrep]{sparse_fr_spectrum}}.}
}
\description{
Creates a plot visualising computations for Hutchinson & Knopoff's model.
}
