% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{hutch_y}
\alias{hutch_y}
\title{Critical bandwidth distance (Hutchison & Knopoff)}
\usage{
hutch_y(f1, f2)
}
\arguments{
\item{f1}{(Numeric vector) Frequency 1, Hz}

\item{f2}{(Numeric vector) Frequency 2, Hz}
}
\value{
(Numeric vector) Unsigned distances in frequency bandwidths.
}
\description{
Calculates the distance between pairs of frequencies in units of
critical bandwidths, according to the model of
\insertCite{Hutchinson1978;textual}{incon}.
}
\references{
\insertAllCited{}
}
