% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\docType{data}
\name{incon_models}
\alias{incon_models}
\title{Simultaneous consonance models}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 17 rows and 8 columns.
}
\usage{
incon_models
}
\description{
This \code{\link[tibble]{tibble}} summarises the
consonance models available in the \code{incon} package.
\itemize{
\item \code{label}: Label for the model, used by \code{\link{incon}}.
\item \code{citation}: Citation for the model.
\item \code{class}: Organises the models by psychological theory.
\item \code{package}: Identifies the package in which the model is implemented.
\item \code{consonance}:
\code{TRUE} if the model should positively correlate with consonance,
\code{FALSE} if the model should negatively correlate with consonance.
\item \code{spectrum_sensitive}: Whether the model is sensitive to
the spectral characteristics of the input,
in particular the number of harmonics
and the roll-off rate.
\item \code{continuous_pitch}: Whether the model can take continuous pitch inputs.
\item \code{f}: Function to call the model.
}
}
\keyword{data}
