% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-jl12.R
\name{jl_rule_1}
\alias{jl_rule_1}
\title{Tonal Dissonance, Rule 1}
\usage{
jl_rule_1(pc_set)
}
\arguments{
\item{pc_set}{(Numeric vector) Pitch-class set to analyse.
No input checking is performed.}
}
\value{
1, 2, or 3, corresponding to increasing degrees of dissonance.
}
\description{
"Chords occurring in a major scale should be less dissonant
than chords occurring only in a minor scale,
which in turn should be less dissonant than chords
occurring in neither sort of scale."
\insertCite{Johnson-Laird2012}{jl12}.
}
\references{
\insertAllCited{}
}
