% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-jl12.R
\name{jl_rule_3}
\alias{jl_rule_3}
\title{Tonal Dissonance, Rule 3}
\usage{
jl_rule_3(pc_set)
}
\arguments{
\item{pc_set}{(Numeric vector) Pitch-class set to analyse.}
}
\value{
\code{FALSE} for consonant, \code{TRUE} for dissonant.
If a consonant solution is found,
the stacked-thirds version of the chord is returned
as the attribute \code{"solution"},
accessible with the command \code{attr(x, "solution")}.
}
\description{
"Chords built from intervals of a third should be more consonant
than chords that are not built from thirds."
"The principle allows for just one missing third intervening
between two pitch classes a fifth apart"
\insertCite{Johnson-Laird2012}{incon}.
}
\references{
\insertAllCited{}
}
