% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-jl12.R
\name{jl_tonal_dissonance}
\alias{jl_tonal_dissonance}
\alias{jl_tonal_dissonance.default}
\alias{jl_tonal_dissonance.pc_set}
\title{Tonal dissonance}
\usage{
jl_tonal_dissonance(x)

\method{jl_tonal_dissonance}{default}(x)

\method{jl_tonal_dissonance}{pc_set}(x)
}
\arguments{
\item{x}{Sonority to analyse.
This will be coerced to an object of class \code{\link[hrep]{pc_set}}.}
}
\value{
Integer scalar identifying the chord's consonance rank,
with higher values corresponding to increasing degrees of dissonance.
}
\description{
Computes tonal dissonance using the algorithm
of \insertCite{Johnson-Laird2012;textual}{jl12}.
}
\examples{
jl_tonal_dissonance(c(0, 4, 7))
jl_tonal_dissonance(c(0, 3, 7))
jl_tonal_dissonance(c(0, 3, 6))
}
\references{
\insertAllCited{}
}
