% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-har18.R
\name{kl_div_from_uniform}
\alias{kl_div_from_uniform}
\alias{kl_div_from_uniform.smooth_spectrum}
\title{Kullback-Leibler divergence from uniform}
\usage{
kl_div_from_uniform(x)

\method{kl_div_from_uniform}{smooth_spectrum}(x)
}
\arguments{
\item{x}{Input distribution.}
}
\value{
The Kullback-Leibler divergence from a uniform distribution
to the input distribution.
}
\description{
Gets the Kullback-Leibler divergence of a provided distribution
from a uniform distribution.
}
