% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{list_models}
\alias{list_models}
\title{List models}
\usage{
list_models()
}
\value{
A character vector of consonance/dissonance models.
}
\description{
Lists the consonance/dissonance models that can be selected in
\code{\link{incon}},
}
