% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{multiplicity}
\alias{multiplicity}
\alias{multiplicity.parn94}
\alias{multiplicity.default}
\title{Get multiplicity}
\usage{
multiplicity(x, k_s = parn94_params()$k_s, ...)

\method{multiplicity}{parn94}(x, k_s = parn94_params()$k_s, ...)

\method{multiplicity}{default}(x, k_s = parn94_params()$k_s, ...)
}
\arguments{
\item{x}{Object to analyse.}

\item{k_s}{Numeric scalar, parameter from Parncutt & Strasburger (1994).}

\item{...}{Further parameters to pass to \code{\link{incon}()}.}
}
\value{
Multiplicity, a numeric scalar.
}
\description{
Computes the multiplicity of a sound, after
\insertCite{Parncutt1994;textual}{incon}.
}
\references{
\insertAllCited{}
}
