% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{parn94}
\alias{parn94}
\alias{parn94.default}
\alias{parn94.sparse_pi_spectrum}
\title{Parncutt & Strasburger (1994)}
\usage{
parn94(x, par = parn94_params(), ...)

\method{parn94}{default}(x, par = parn94_params(), ...)

\method{parn94}{sparse_pi_spectrum}(x, par = parn94_params(), ...)
}
\arguments{
\item{x}{Object to analyse,
which will be coerced to an object of class
\code{\link[hrep]{sparse_pi_spectrum}}.
Various input types are possible:
\itemize{
\item Numeric vectors will be treated as vectors of MIDI note numbers,
which will be expanded into their implied harmonics.
\item A two-element list can be used to define a harmonic spectrum.
The first element should be a vector of MIDI note numbers,
the second a vector of amplitudes.
\item The function also accepts classes from the \code{hrep} package,
such as produced by \code{\link[hrep]{pi_chord}()} and
\code{\link[hrep]{sparse_pi_spectrum}()}.
}}

\item{par}{Parameter list as created by \code{\link{parn94_params}()}.}

\item{...}{Parameters to pass to \code{\link[hrep]{sparse_pi_spectrum}}.
\itemize{
\item \code{num_harmonics}: Number of harmonics to use when expanding
chord tones into their implied harmonics.
\item \code{roll_off}: Rate of amplitude roll-off for the harmonics.
}}
}
\value{
An list of class \code{parn94}, comprising the following components:
\item{pure_spectrum}{A tibble describing the sonority's pure spectrum.
The pure spectrum is a spectral representation of the input sound
after auditory masking, but before pattern matching.}
\item{pure_spectrum}{A tibble describing the sonority's complex spectrum.
The complex spectrum is created from the pure spectrum through
harmonic template matching.}
\item{combined_spectrum}{A tibble describing the sonority's combined spectrum.
The combined spectrum corresponds to the combination of the
pure and complex spectra.}
\item{par}{A list comprising the parameters used to perform the analysis,
as created by \code{\link{parn94_params}()}.}
}
\description{
This function analyses a sonority using Richard Parncutt's
psychoacoustic model of harmony, as described in
\insertCite{Parncutt1994;textual}{parn94}.
}
\references{
\insertAllCited{}
}
