% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{parn94_params}
\alias{parn94_params}
\title{Model parameters}
\usage{
parn94_params(
  unit_amplitude_in_dB = 60,
  template_num_harmonics = 11,
  template_roll_off = 1,
  k_t = 3,
  k_p = 0.5,
  k_c = 0.2,
  k_s = 0.5,
  k_m = 12,
  al_0 = 15,
  min_midi = 0,
  max_midi = 120
)
}
\arguments{
\item{unit_amplitude_in_dB}{(Numeric scalar) Describes the number of decibels
to assign to a spectral component of amplitude 1.
By default, amplitude 1 corresponds to the amplitude of each chord pitch's
fundamental frequency (though this can be changed by expressing input chords
as pitch-class sparse spectra, see \code{\link[hrep]{sparse_pi_spectrum}()}).}

\item{template_num_harmonics}{(Integerish scalar) Number of harmonics to
include in the spectral template, including the fundamental frequency.}

\item{template_roll_off}{(Numeric scalar) Roll-off rate for the
harmonic template. This parameter is passed to
\code{\link[hrep]{sparse_pi_spectrum}()}.}

\item{k_t}{(Numeric scalar) Relative perceptual weighting of complex versus pure tones.}

\item{k_p}{(Numeric scalar) Scaling coefficient for pure sonorousness.}

\item{k_c}{(Numeric scalar) Scaling coefficient for complex sonorousness.}

\item{k_s}{(Numeric scalar) Scaling coefficient for multiplicity.}

\item{k_m}{(Numeric scalar) Gradient of a pure tone's masking pattern.}

\item{al_0}{(Numeric scalar) Audibility saturation rate.}

\item{min_midi}{(Numeric scalar) Lowest MIDI pitch considered by the model.}

\item{max_midi}{(Numeric scalar) Highest MIDI pitch considered by the model.}
}
\value{
A list of parameters which can be passed to analysis functions
such as \code{\link{parn94}}.
}
\description{
This function compiles parameters for Parncutt's psychoacoustic model.
The parameters are defined with reference to
\insertCite{Parncutt1994;textual}{incon}.
}
\references{
\insertAllCited{}
}
