% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{pitch_commonality}
\alias{pitch_commonality}
\title{Get pitch commonality}
\usage{
pitch_commonality(x, y, ...)
}
\arguments{
\item{x}{The first sonority to compare, passed to \code{\link{pitch_salience}()}.
Typically will be a numeric vector of MIDI pitches.}

\item{y}{The second sonority to compare, passed to \code{\link{pitch_salience}()}.
Typically will be a numeric vector of MIDI pitches.}

\item{...}{Further arguments to pass to \code{\link{pitch_salience}()}.}
}
\value{
Pitch commonality, as a numeric scalar.
}
\description{
Gets the pitch commonality between two sonorities, after
\insertCite{Parncutt1994;textual}{incon}.
}
\references{
\insertAllCited{}
}
