% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{pitch_salience}
\alias{pitch_salience}
\alias{pitch_salience.default}
\alias{pitch_salience.pitch_salience}
\alias{pitch_salience.parn94}
\title{Get pitch salience}
\usage{
pitch_salience(x, ...)

\method{pitch_salience}{default}(x, ...)

\method{pitch_salience}{pitch_salience}(x, ...)

\method{pitch_salience}{parn94}(x, ...)
}
\arguments{
\item{x}{Object to analyse, passed to \code{\link{parn94}()}.}

\item{...}{Further arguments to pass to \code{\link{parn94}()}.}
}
\value{
Returns a vector where each element describes
the salience of a different chromatic pitch.
The first element of this vector corresponds to the
\code{min_midi} argument from \code{\link{parn94_params}},
and the last element corresponds to the \code{max_midi} argument.
}
\description{
Analyses the pitch salience of a sonority, after
\insertCite{Parncutt1994;textual}{incon}.
}
\references{
\insertAllCited{}
}
