% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-corpdiss.R
\docType{data}
\name{popular_1_pc_chord_type}
\alias{popular_1_pc_chord_type}
\title{Consonance tables for popular music}
\description{
This table summarises chord prevalences in the
McGill Billboard corpus \insertCite{Burgoyne2011}{hcorp}.
These pieces were sampled from the Billboard magazine's
United States "Hot 100" chart between 1958 and 1991,
and transcribed by expert musicians.
See `?hcorp::popular_1` for more details.
}
\details{
Chords are represented as pitch-class chord types:
see \code{\link[hrep]{pc_chord_type}} for details.
}
\references{
\insertAllCited{}
}
\keyword{data}
