% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn94.R
\name{pure_sonor}
\alias{pure_sonor}
\alias{pure_sonor.parn94}
\alias{pure_sonor.default}
\title{Get pure sonorousness}
\usage{
pure_sonor(x, k_p = parn94_params()$k_p, ...)

\method{pure_sonor}{parn94}(x, k_p = parn94_params()$k_p, ...)

\method{pure_sonor}{default}(x, k_p = parn94_params()$k_p, ...)
}
\arguments{
\item{x}{Object to analyse.}

\item{k_p}{Parncutt & Strasburger (1994) set this to 0.5 (p. 105).}

\item{...}{Further parameters to pass to \code{\link{parn94}()}.}
}
\value{
Pure sonorousness, a numeric scalar.
}
\description{
Computes the pure sonorousness of a sound, after
\insertCite{Parncutt1994;textual}{parn94}.
}
\references{
\insertAllCited{}
}
