% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-bowl18.R
\docType{data}
\name{rational_scale}
\alias{rational_scale}
\title{Rational scale}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 2 rows and 12 columns.
}
\usage{
rational_scale
}
\description{
This defines the rational scale used when computing harmonicity with
\code{\link{gill09_harmonicity}}.
It is a matrix with 2 rows and 12 columns,
where the first row corresponds to fraction numerators,
and the second row corresponds to fraction denominators.
Column i identifes the interval of size (i - 1) semitones.
For example, column 8 identifies the perfect fifth
(7 semitones) as a 3:2 ratio.
}
\keyword{data}
