% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn88.R
\name{root}
\alias{root}
\title{Root}
\usage{
root(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{parn88}}.}
}
\value{
The estimated chord root (integer scalar).
}
\description{
Estimates the chord root of a pitch-class set using the root-finding model of
\insertCite{Parncutt1988;textual}{parn88}.
This function is a wrapper for \code{\link{parn88}}.
}
\references{
\insertAllCited{}
}
