% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn88.R
\name{root_ambiguity}
\alias{root_ambiguity}
\title{Root ambiguity}
\usage{
root_ambiguity(...)
}
\arguments{
\item{...}{Arguments to pass to \code{\link{parn88}}.}
}
\value{
The root ambiguity (numeric scalar).
}
\description{
Estimates the root ambiguity of a pitch-class set using the root-finding model of
\insertCite{Parncutt1988;textual}{incon}.
This function is a wrapper for \code{\link{parn88}}.
}
\references{
\insertAllCited{}
}
