% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn88.R
\docType{data}
\name{root_by_pc_chord}
\alias{root_by_pc_chord}
\title{Root by pitch-class chord}
\description{
This vector stores precomputed chord roots for
all pitch-class chords,
indexed by their encoding as defined in the "hrep" package.
See \code{\link[hrep]{pc_chord}} for more details.
}
\keyword{data}
