% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-parn88.R
\docType{data}
\name{root_support_weights}
\alias{root_support_weights}
\title{Root support weights}
\format{
An object of class \code{list} of length 2.
}
\usage{
root_support_weights
}
\description{
A list of different root support weights that may be used
by the root-finding algorithm of \insertCite{Parncutt1988;textual}{incon}.
See \code{\link{parn88}} for more information.
}
\references{
\insertAllCited{}
}
\keyword{datasets}
