% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{roughness_hutch}
\alias{roughness_hutch}
\alias{roughness_hutch.default}
\alias{roughness_hutch.sparse_fr_spectrum}
\title{Spectral roughness (Hutchinson & Knopoff)}
\usage{
roughness_hutch(
  x,
  cbw_cut_off = 1.2,
  a = 0.25,
  b = 2,
  dissonance_function = hutch_dissonance_function,
  ...
)

\method{roughness_hutch}{default}(
  x,
  cbw_cut_off = 1.2,
  a = 0.25,
  b = 2,
  dissonance_function = hutch_dissonance_function,
  ...
)

\method{roughness_hutch}{sparse_fr_spectrum}(
  x,
  cbw_cut_off = 1.2,
  a = 0.25,
  b = 2,
  dissonance_function = hutch_dissonance_function,
  ...
)
}
\arguments{
\item{x}{Object to analyse, which is coerced to the class
\code{\link[hrep]{sparse_fr_spectrum}}.
\itemize{
\item Numeric vectors will be treated as vectors of MIDI note numbers,
and expanded into their implied harmonics.
\item Two-element lists will be treated as finalised spectra,
with the first element being a numeric vector of frequencies,
and the second element being a numeric vector of amplitudes.
}}

\item{cbw_cut_off}{Parameter passed to \code{\link{hutch_g}()}.}

\item{a}{Parameter passed to \code{\link{hutch_g}()}.}

\item{b}{Parameter passed to \code{\link{hutch_g}()}.}

\item{dissonance_function}{Function for computing dissonance contribution as a function of
critical bandwidth distance, defaulting to \code{\link{hutch_dissonance_function}}.
Custom functions may be specified here as long as they use the same parameter list
as the original function.}

\item{...}{Further arguments to pass to \code{\link[hrep]{sparse_fr_spectrum}}.}
}
\value{
Numeric scalar, identifying the roughness of the spectrum.
}
\description{
Gets the roughness of a sonority according to the model of
\insertCite{Hutchinson1978;textual}{incon}.
}
\references{
\insertAllCited{}
}
