% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-dycon.R
\name{roughness_vass}
\alias{roughness_vass}
\alias{roughness_vass.default}
\alias{roughness_vass.sparse_fr_spectrum}
\title{Spectral roughness (Vassilakis)}
\usage{
roughness_vass(x, ...)

\method{roughness_vass}{default}(x, ...)

\method{roughness_vass}{sparse_fr_spectrum}(x, ...)
}
\arguments{
\item{x}{Object to analyse, which is coerced to the class
\code{\link[hrep]{sparse_fr_spectrum}}.
\itemize{
\item Numeric vectors will be treated as vectors of MIDI note numbers,
and expanded into their implied harmonics.
\item Two-element lists will be treated as finalised spectra,
with the first element being a numeric vector of frequencies,
and the second element being a numeric vector of amplitudes.
}}

\item{...}{Further arguments to pass to \code{\link[hrep]{sparse_fr_spectrum}}.}
}
\value{
Estimated roughness, as a numeric scalar.
}
\description{
Gets the roughness of a sonority according to the model of
\insertCite{Vassilakis2001;textual}{incon}
\insertCite{Villegas2010;textual}{incon}
}
\references{
\insertAllCited{}
}
