% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-stolz15.R
\name{smooth_log_periodicity}
\alias{smooth_log_periodicity}
\alias{smooth_log_periodicity.default}
\alias{smooth_log_periodicity.pi_chord}
\title{Smoothed log periodicity}
\usage{
smooth_log_periodicity(x, d = 0.011)

\method{smooth_log_periodicity}{default}(x, d = 0.011)

\method{smooth_log_periodicity}{pi_chord}(x, d = 0.011)
}
\arguments{
\item{x}{Sonority to analyse.
This will be coerced to an object of class \code{\link[hrep]{pi_chord}}.
Numeric inputs will be interpreted as MIDI note numbers.}

\item{d}{(numeric scalar):
Maximal allowed error in the algorithm's
interval approximation step, expressed as
a fraction of the original interval.
The default value, 0.011, corresponds to 'Rational Tuning II'
in Stolzenburg's paper.}
}
\value{
A numeric scalar identifying the chord's periodicity.
High values mean a higher period length, lower periodicity,
and lower consonance.
}
\description{
This function computes a chord's smoothed logarithmic periodicity,
after \insertCite{Stolzenburg2015;textual}{incon}.
}
\references{
\insertAllCited{}
}
