% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-har18.R
\name{sweep_harmonic_template}
\alias{sweep_harmonic_template}
\alias{sweep_harmonic_template.pc_set}
\alias{sweep_harmonic_template.milne_pc_spectrum}
\title{Sweep harmonic template}
\usage{
sweep_harmonic_template(
  x,
  num_harmonics = 12,
  rho = 0.75,
  sigma = 6.83,
  array_dim = 1200,
  ...
)

\method{sweep_harmonic_template}{pc_set}(
  x,
  num_harmonics = 12,
  rho = 0.75,
  sigma = 6.83,
  array_dim = 1200,
  ...
)

\method{sweep_harmonic_template}{milne_pc_spectrum}(x, num_harmonics = 12, rho = 0.75, sigma = 6.83, ...)
}
\arguments{
\item{x}{Object to analyse.}

\item{num_harmonics}{See \code{\link{pc_harmonicity}}.}

\item{rho}{See \code{\link{pc_harmonicity}}.}

\item{sigma}{See \code{\link{pc_harmonicity}}.}

\item{array_dim}{See \code{\link{pc_harmonicity}}.}

\item{...}{Arguments passed to specific methods.}
}
\value{
An object of class \code{\link[hrep]{milne_pc_spectrum}},
identifying each pitch class with a perceptual weight
corresponding to its harmonic template fit.
}
\description{
Sweeps a harmonic template over an input spectrum.
}
