% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model-har18.R
\name{sweep_template}
\alias{sweep_template}
\title{Sweep template}
\usage{
sweep_template(x, template, legacy = FALSE)
}
\arguments{
\item{x}{(Numeric vector)
The vector to be swept over.}

\item{template}{(Numeric vector)
The template to sweep over \code{x}.
Should have the same dimensionality as \code{x}.}

\item{legacy}{(Logical scalar)
Whether to use the legacy R implementation
(default = \code{FALSE}).
Otherwise the faster C++ implementation is used.}
}
\description{
Sweeps a circular template over a circular vector
and computes the cosine similarity at each possible offset.
}
