% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/top-level.R
\name{incon}
\alias{incon}
\title{Instantaneous consonance}
\usage{
incon(x, model = "hutch_78_roughness", x_int = NULL,
  num_harmonics = 11L, roll_off = 1, par = list())
}
\arguments{
\item{x}{Chord to analyse.
This will be coerced to an object of class \code{\link[hrep]{pi_chord}}.
\itemize{
\item Numeric vectors will be interpreted as vectors of MIDI note numbers.
\item Character scalars of the form \code{"60 64 67"} are also accepted;
these will likewise be interpreted as MIDI note numbers.
}}

\item{model}{(Character vector, default = "hutch_78_roughness")
Computational model(s) to apply.
See \strong{Details} for available values.
Alternatively, \code{"all"} will select all models.}

\item{x_int}{While some models accept continuous pitch values as inputs,
others do not (see \code{\link{incon_models}} for details).
Correspondingly, the user may specify one chord (\code{x})
to deliver to continuous-pitch models,
and another chord (\code{x_int}) to deliver to non-continuous-pitch models.
If \code{x_int} is \code{NULL},
then non-continuous-pitch models will default to analysing \code{x},
and will throw an error if \code{x} is non-integer.}

\item{num_harmonics}{(Integerish scalar, default = 11)
Number of harmonics to which chord tones should be expanded,
including the fundamental frequency.
This only affects a subset of the consonance models.}

\item{roll_off}{(Integerish scalar, default = 1)
Roll-off rate for the upper harmonics in complex tones.
If the roll-off rate is \eqn{x},
then the amplitude of the ith harmonic is equal to \eqn{1 ^ -x}.}

\item{par}{(List, default = \code{list()})
Optional list containing additional parameters to pass to specific models.
If non-empty, this list should be a named list of lists,
with each name identifying the model to which additional parameters
should be passed (see \code{\link{list_models}}) for legal names.
Each element should itself be a named list of parameters,
with the names identifying the arguments,
and the values being those to be passed to the model function.}
}
\value{
A named numeric vector comprising the outputs of each computational model
in the order that they were specified.
}
\description{
Computes the instantaneous consonance (or dissonance) of a sonority
according to various computational models.
}
\details{
The following models are available:
\itemize{
\item \code{gill_09_harmonicity}:
the harmonicity model of \insertCite{Gill2009;textual}{bowl18}
(see \code{bowl18::\link[bowl18]{gill09_harmonicity}}).
\item \code{har_18_harmonicity}:
the harmonicity model of \insertCite{Harrison2018;textual}{har18}
(see \code{har18::\link[har18]{pc_harmonicity}}).
\item \code{milne_13_harmonicity}:
the harmonicity model of \insertCite{Milne2013;textual}{har18}
(see \code{har18::\link[har18]{pc_harmonicity}}).
\item \code{parn_88_root_ambig}:
the root ambiguity model of \insertCite{Parncutt1988;textual}{parn88}
(see \code{parn88::\link[parn88]{root_ambiguity}}).
\item \code{parn_94_complex}:
the complex sonorousness feature of \insertCite{Parncutt1994;textual}{parn94}
(see \code{parn94::\link[parn94]{complex_sonor}}).
\item \code{stolz_15_periodicity}:
smoothed logarithmic periodicity,
after \insertCite{Stolzenburg2015;textual}{stolz15}
(see \code{stolz15::\link[stolz15]{smooth_log_periodicity}}).
\item \code{bowl_18_min_freq_dist}:
the minimum frequency distance feature of
\insertCite{Bowling2018;textual}{bowl18}
(see \code{bowl18::\link[bowl18]{bowl18_min_freq_dist}}).
\item \code{huron_94_dyadic}:
aggregate dyadic consonance, after \insertCite{Huron1994;textual}{incon}.
\item \code{hutch_78_roughness}:
the roughness model of \insertCite{Hutchinson1978;textual}{dycon}
(see \code{dycon::\link[dycon]{roughness_hutch}}).
\item \code{parn_94_pure}:
the complex sonorousness feature of \insertCite{Parncutt1994;textual}{parn94}
(see \code{parn94::\link[parn94]{pure_sonor}}).
\item \code{seth_93_roughness}:
the roughness model of \insertCite{Sethares1993;textual}{dycon}
(see \code{dycon::\link[dycon]{roughness_seth}}).
\item \code{vass_01_roughness}:
the roughness model of \insertCite{Vassilakis2001;textual}{dycon}
(see \code{dycon::\link[dycon]{roughness_vass}}).
\item \code{wang_13_roughness}:
the roughness model of \insertCite{Wang2013;textual}{wang13}
(see \code{wang13::\link[wang13]{roughness_wang}}).
\item \code{jl_12_tonal}:
the tonal dissonance model of \insertCite{Johnson-Laird2012;textual}{jl12}
(see \code{jl12::\link[jl12]{jl_tonal_dissonance}}).
\item \code{har_19_corpus}:
a corpus-based model of cultural familiarity
\insertCite{Harrison2019}{incon}
(see \code{corpdiss::\link[corpdiss]{corpus_dissonance}}).
\item \code{parn_94_mult}:
the multiplicity feature of \insertCite{Parncutt1994;textual}{parn94}
(see \code{parn94::\link[parn94]{multiplicity}}).
\item \code{har_19_composite}:
a model combining spectral interference \insertCite{Hutchinson1978}{dycon},
periodicity \insertCite{Harrison2018}{har18},
and cultural familiarity,
as introduced by \insertCite{Harrison2019;textual}{incon}.
}
}
\examples{
incon(c(60, 64, 67))
incon("60 64 67")
incon("60 64 67", model = "hutch_78_roughness")
incon("60 64 67", model = c("hutch_78_roughness", "parn_88_root_ambig"))
incon("0 4 7", model = "gill_09_harmonicity",
      par = list(gill_09_harmonicity = list(tonic = 3)))

}
\references{
\insertAllCited{}
}
